+++ client/src/components/TrailList.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { FlatList } from 'react-native';
 import { ListItem } from 'react-native-elements'
-import { connect } from 'react-redux';
 
 class TrailList extends Component {
     render() {
